#1554 Split VideoCard to separate component
